requirements.txt
Django~=4.0.2
mysqlclient
django-mysql
djangorestframework~=3.13.1